Move Overwatch services to its own module.
